Thermochron

Documentation for the Thermochron.jl package.

Thermochron.ApatiteHe

Thermochron.PlanarAr

Thermochron.PlanarHe

Thermochron.SphericalAr

Thermochron.SphericalHe

Thermochron.ZirconHe

StatGeochemBase.image_from_paths

Thermochron.MCMC

Thermochron.MCMC_varkinetics

Thermochron.anneal

Thermochron.anneal!

Thermochron.chronometers

Thermochron.lcmod

Thermochron.modelage

Thermochron.modelage

Thermochron.modelage

Thermochron.modellength

Thermochron.reltrackdensity

Thermochron.rmr0fromdpar

Thermochron.rmr0model

Thermochron.simannealT

Thermochron.slabsphereintersectiondensity

Thermochron.slabsphereintersectionfraction

Thermochron.sphereintersectiondensity

Thermochron.sphereintersectionfraction

Thermochron.ApatiteHe — Type

ApatiteHe(T=Float64;
    age::Number=T(NaN),
    age_sigma::Number=T(NaN),
    offset=zero(T),
    r::Number, 
    dr::Number=one(T), 
    U238::Number, 
    Th232::Number, 
    Sm147::Number=zero(T), 
    U238_matrix::Number=zero(T), 
    Th232_matrix::Number=zero(T), 
    Sm147_matrix::Number=zero(T), 
    agesteps::AbstractRange,
)

Construct an ApatiteHe chronometer representing an apatite with a raw helium age of age ± age_sigma [Ma], a radius of r [μm], and uniform U, Th and Sm concentrations specified by U238, Th232, and Sm147 [PPMW]. A present day U-235/U-238 ratio of 1/137.818 is assumed.

Spatial discretization follows a radius step of dr [μm], and temporal discretization follows the age steps specified by the agesteps range, in Ma.

Thermochron.PlanarAr — Type

PlanarAr(T=Float64;
    age::Number=T(NaN),
    age_sigma::Number=T(NaN),
    offset=zero(T),
    D0::Number,
    Ea::Number,
    r::Number, 
    dr::Number=one(T), 
    K40::Number, 
    agesteps::AbstractRange,
)

Construct an PlanarAr chronometer representing a mineral with a raw argon age of age ± age_sigma [Ma], a uniform diffusivity specified by a frequency factor D0 [cm^2/s] and an activation energy Ea [kJ/mol], a radius of r [μm], and uniform K-40 concentrations specified by K40 [PPM].

Spatial discretization follows a halfwidth step of dr [μm], and temporal discretization follows the age steps specified by the agesteps range, in Ma.

Thermochron.PlanarHe — Type

PlanarHe(T=Float64;
    age::Number=T(NaN),
    age_sigma::Number=T(NaN),
    offset=zero(T),
    D0::Number,
    Ea::Number,
    stoppingpower::Number=T(1.189),
    r::Number, 
    dr::Number=one(T), 
    U238::Number, 
    Th232::Number, 
    Sm147::Number=zero(T), 
    U238_matrix::Number=zero(T), 
    Th232_matrix::Number=zero(T), 
    Sm147_matrix::Number=zero(T), 
    agesteps::AbstractRange,
)

Construct an PlanarHe chronometer representing a mineral with a raw helium age of age ± age_sigma [Ma], a uniform diffusivity specified by a frequency factor D0 [cm^2/sec] and an activation energy Ea [kJ/mol], a halfwidth of r [μm], and uniform U, Th and Sm concentrations specified by U238, Th232, and Sm147 [PPM]. (A present day U-235/U-238 ratio of 1/137.818 is assumed)

Spatial discretization follows a halfwidth step of dr [μm], and temporal discretization follows the age steps specified by the agesteps range, in Ma.

Thermochron.SphericalAr — Type

SphericalAr(T=Float64;
    age::Number=T(NaN),
    age_sigma::Number=T(NaN),
    offset=zero(T),
    D0::Number,
    Ea::Number,
    r::Number, 
    dr::Number=one(T), 
    K40::Number, 
    agesteps::AbstractRange,
)

Construct an SphericalAr chronometer representing a mineral with a raw argon age of age ± age_sigma [Ma], a uniform diffusivity specified by a frequency factor D0 [cm^2/s] and an activation energy Ea [kJ/mol], a radius of r [μm], and uniform K-40 concentrations specified by K40 [PPM].

Spatial discretization follows a radius step of dr [μm], and temporal discretization follows the age steps specified by the agesteps range, in Ma.

Thermochron.SphericalHe — Type

SphericalHe(T=Float64;
    age::Number=T(NaN),
    age_sigma::Number=T(NaN),
    offset=zero(T),
    D0::Number,
    Ea::Number,
    stoppingpower::Number=T(1.189),
    r::Number, 
    dr::Number=one(T), 
    U238::Number, 
    Th232::Number, 
    Sm147::Number=zero(T), 
    U238_matrix::Number=zero(T), 
    Th232_matrix::Number=zero(T), 
    Sm147_matrix::Number=zero(T), 
    agesteps::AbstractRange,
)

Construct an SphericalHe chronometer representing a mineral with a raw helium age of age ± age_sigma [Ma], a uniform diffusivity specified by a frequency factor D0 [cm^2/sec] and an activation energy Ea [kJ/mol], a radius of r [μm], and uniform U, Th and Sm concentrations specified by U238, Th232, and Sm147 [PPM]. (A present day U-235/U-238 ratio of 1/137.818 is assumed)

Spatial discretization follows a radius step of dr [μm], and temporal discretization follows the age steps specified by the agesteps range, in Ma.

Thermochron.ZirconHe — Type

ZirconHe(T=Float64;
    age::Number=T(NaN),
    age_sigma::Number=T(NaN),
    T(offset),
    r::Number=one(T),
    dr::Number, 
    U238::Number,
    Th232::Number,
    Sm147::Number=zero(T),
    U238_matrix::Number=zero(T), 
    Th232_matrix::Number=zero(T), 
    Sm147_matrix::Number=zero(T), 
    agesteps::AbstractRange
)

Construct a ZirconHe chronometer representing a zircon with a raw helium age of age ± age_sigma [Ma], a radius of r [μm], and uniform U, Th and Sm concentrations specified by U238, Th232, and Sm147 [PPMw]. A present day U-235/U-238 ratio of 1/137.818 is assumed.

Spatial discretization follows a radius step of dr μm, and temporal discretization follows the age steps specified by the agesteps range, in Ma.

StatGeochemBase.image_from_paths — Method

image_from_paths(tT::TTResult; 
    xresolution::Int=1800, 
    yresolution::Int=1200, 
    xrange = nanextrema(tT.tpointdist), 
    yrange = nanextrema(tT.Tpointdist), 
)

Produce a 2d image (histogram) of path densities given a TTResult

julia> imgcounts, xq, yq = image_from_paths(tT; xrange=boundary.agepoints, yrange=boundary.T₀)

Thermochron.MCMC — Method

MCMC(data::Vector{<:Chronometer}, model::NamedTuple, npoints::Int, path.agepoints::Vector, path.Tpoints::Vector, constraint::Constraint, boundary::Boundary, [detail::DetailInterval])

Markov chain Monte Carlo time-Temperature inversion of the thermochronometric data specified as a vector chrons of Chronometer objects (ZirconHe, ApatiteHe, ApatiteFT, etc.) and model parameters specified by the named tuple model, with variable diffusion kinetics.

Returns a TTResult object containing posterior time-temperature paths,

See also MCMC_varkinetics for a variant with variable diffusion kinetics.

Examples

tT = MCMC(data::NamedTuple, model::NamedTuple, constraint::Constraint, boundary::Boundary, [detail::DetailInterval])

Thermochron.MCMC_varkinetics — Method

MCMC_varkinetics(chrons::Vector{<:Chronometer}, model::NamedTuple, npoints::Int, path.agepoints::Vector, path.Tpoints::Vector, constraint::Constraint, boundary::Boundary, [detail::DetailInterval])

Markov chain Monte Carlo time-Temperature inversion of the thermochronometric data specified as a vector chrons of Chronometer objects (ZirconHe, ApatiteHe, ApatiteFT, etc.) and model parameters specified by the named tuple model, with variable diffusion kinetics.

Returns a TTResult object containing posterior time-temperature paths, and a KineticResult object containing the posterior kinetic parameters.

See also MCMC for a variant with constant diffusion kinetics.

Examples

tT, kinetics = MCMC_varkinetics(data::NamedTuple, model::NamedTuple, constraint::Constraint, boundary::Boundary, [detail::DetailInterval])

Thermochron.anneal — Function

ρᵣ = anneal(dt::Number, tsteps::Vector, Tsteps::Matrix, [model::DiffusivityModel=ZRDAAM()])

ZirconHe damage annealing model as in Guenthner et al. 2013 (AJS)

Thermochron.anneal! — Method

anneal!(data::Vector{<:Chronometer}, ::Type{<:HeliumSample}, tsteps, Tsteps, dm::DiffusivityModel)
anneal!(mineral::ZirconHe, Tsteps, dm::ZRDAAM)
anneal!(mineral::ApatiteHe, Tsteps, dm::RDAAM)
anneal!(ρᵣ::Matrix, dt::Number, tsteps, Tsteps, [dm::DiffusivityModel=ZRDAAM()])

In-place version of anneal

Thermochron.chronometers — Method

chronometers([T=Float64], data, model)

Construct a vector of Chronometer objects given a dataset data and model parameters model

Thermochron.lcmod — Method

lcmod(l, θ)

Calculate the model c-axis equivalent length ("lc,mod") given a measured "confined" fission track length l [microns] and angle from the c-axis θ [degrees] following the approach of Donelick et al. 1999 (doi: 10.2138/am-1999-0902)

Thermochron.modelage — Method

modelage(mineral::ZirconHe, Tsteps, [ρᵣ], dm::ZRDAAM)
modelage(mineral::ApatiteHe, Tsteps, [ρᵣ], dm::RDAAM)
modelage(mineral::SphericalHe, Tsteps)
modelage(mineral::PlanarHe, Tsteps)

Calculate the predicted bulk U-Th/He age of a zircon, apatite, or other mineral that has experienced a given t-T path (specified by mineral.tsteps for time and Tsteps for temperature, at a time resolution of step(mineral.tsteps)) using a Crank-Nicolson diffusion solution for a spherical (or planar slab) grain of radius (or halfwidth) mineral.r at spatial resolution mineral.dr.

Spherical implementation based on the the Crank-Nicolson solution for diffusion out of a spherical mineral crystal in Ketcham, 2005 (doi: 10.2138/rmg.2005.58.11).

Thermochron.modelage — Method

modelage(mineral::SphericalAr, Tsteps)
modelage(mineral::PlanarAr, Tsteps)

Calculate the precdicted bulk K/Ar age of a mineral that has experienced a given t-T path (specified by mineral.tsteps for time and Tsteps for temperature, at a time resolution of step(mineral.tsteps)) using a Crank-Nicholson diffusion solution for a spherical (or planar slab) grain of radius (or halfwidth ) mineral.r at spatial resolution mineral.dr.

Spherical implementation based on the the Crank-Nicolson solution for diffusion out of a spherical mineral crystal in Ketcham, 2005 (doi: 10.2138/rmg.2005.58.11).

Thermochron.modelage — Method

modelage(mineral::ZirconFT, Tsteps, am::ZirconAnnealingModel)
modelage(mineral::MonaziteFT, Tsteps, am::MonaziteAnnealingModel)
modelage(mineral::ApatiteFT, Tsteps, am::ApatiteAnnealingModel)

Calculate the precdicted fission track age of an apatite that has experienced a given t-T path (specified by mineral.tsteps for time and Tsteps for temperature, at a time resolution of step(mineral.tsteps)) and given annealing model parameters am.

Possible annealing model types and the references for the equations which they respetively implement include Ketcham1999FC Fanning Curvilinear apatite model of Ketcham et al. 1999 (doi: 10.2138/am-1999-0903) Ketcham2007FC Fanning Curvilinear apatite model of Ketcham et al. 2007 (doi: 10.2138/am.2007.2281) Yamada2007PC Parallel Curvilinear zircon model of Yamada et al. 2007 (doi: 10.1016/j.chemgeo.2006.09.002) Guenthner2013FC Fanning Curvilinear zircon model of Guenthner et al. 2013 (doi: 10.2475/03.2013.01) Jones2021FA Fanning Arrhenius (Fanning Linear) model adapted from Jones et al. 2021 (doi: 10.5194/gchron-3-89-2021)

Thermochron.modellength — Method

modellength(track::ApatiteTrackLength, Tsteps, am::ApatiteAnnealingModel)

Calculate the predicted mean and standard deviation of the distribution of fission track lengths of an apatite that has experienced a given t-T path (specified by track.tsteps for time and Tsteps for temperature, at a time resolution of step(mineral.tsteps)) and given annealing model parameters am.

Possible annealing model types and the references for the equations which they respetively implement include Ketcham1999FC Fanning Curvilinear apatite model of Ketcham et al. 1999 (doi: 10.2138/am-1999-0903) Ketcham2007FC Fanning Curvilinear apatite model of Ketcham et al. 2007 (doi: 10.2138/am.2007.2281)

Thermochron.reltrackdensity — Method

reltrackdensity(t, T, am::AnnealingModel)

Calculate the relative track density ρ corresponding to a given relative track length r

Follows the relations of Ketcham et al. (2000), equations 7a and 7b for apatite and Tagami et al. (1999) for zircon

See also: reltracklength.

Thermochron.rmr0fromdpar — Method

rmr0fromdpar(dpar)

Calculate rmr0 as a function of dpar for "multikinetic" apatite fission track following the relation (Fig. 7) of Ketcham et al. 1999 (doi: 10.2138/am-1999-0903)

rmr0 = 1 - exp(0.647(dpar-1.75) - 1.834)

Thermochron.rmr0model — Function

rmr0model(F, Cl, OH, Mn=0, Fe=0, others=0)

Calculate rmr0 as a function of composition (specified in terms of atoms per fomula unit, or APFU) for "multikinetic" apatite fission track thermochronology.

Implements equation 11 from Ketcham et al. 2007 (doi: 10.2138/am.2007.2281)

rmr0 = (-0.0495 -0.0348F +0.3528|Cl - 1| +0.0701|OH - 1| 
        -0.8592Mn -1.2252Fe -0.1721Others)^0.1433

Thermochron.simannealT — Method

simannealT(n::Integer, Tₐ::Number, λₐ::Number)

To avoid getting stuck in local optima, increase the probability of accepting new proposals at higher annealing "temperature"

Thermochron.slabsphereintersectiondensity — Method

slabsphereintersectiondensity(redges::Vector, ralpha, d)

Calculate the fractional intersection density of an alpha stopping sphere of radius ralpha with each concentric slab-shell (with edges redges and relative volumes rvolumes) of a planar slab crystal where the two are separated by distance d

Thermochron.slabsphereintersectionfraction — Method

slabsphereintersectionfraction(rₚ, rₛ, d)

Calculate the fraction of the surface area of a sphere s with radius rₛ that intersects the interior of a planar slab p of halfwidth rₚ if the two are separated by distance d.

Thermochron.sphereintersectiondensity — Method

sphereintersectiondensity(redges::Vector, rvolumes::Vector, ralpha, d)

Calculate the volume-nomalized fractional intersection density of an alpha stopping sphere of radius ralpha with each concentric shell (with shell edges redges and relative volumes rvolumes) of a spherical crystal where the two are separated by distance d

Thermochron.sphereintersectionfraction — Method

sphereintersectionfraction(r₁, r₂, d)

Calculate the fraction of the surface area of a sphere s₂ with radius r₂ that intersects the interior of a sphere s₁ of radius r₁ if the two are separated by distance d.